uwp: Add conditionals to CustomCode
Exclude operator overloads
Use sealed class instead of struct
Extract extensions code to separate files
