+++ UnitsNet/CustomCode/UnitClasses/Angle.extra.cs
@@ -19,6 +19,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+// Operator overloads not supported in Universal Windows Platform (WinRT Components)
+#if !WINDOWS_UWP
 using System;
 
 namespace UnitsNet
@@ -36,3 +38,4 @@ namespace UnitsNet
         }
     }
 }
+#endif
