+++ UnitsNet/CustomCode/UnitClasses/KinematicViscosity.extra.cs
@@ -19,6 +19,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+// Operator overloads not supported in Universal Windows Platform (WinRT Components)
+#if !WINDOWS_UWP
 using System;
 
 namespace UnitsNet
@@ -56,3 +58,4 @@ namespace UnitsNet
         }
     }
 }
+#endif
