+++ UnitsNet/CustomCode/UnitClasses/SpecificEnergy.extra.cs
@@ -19,6 +19,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+// Operator overloads not supported in Universal Windows Platform (WinRT Components)
+#if !WINDOWS_UWP
 namespace UnitsNet
 {
     public partial struct SpecificEnergy
@@ -49,3 +51,4 @@ namespace UnitsNet
         }
     }
 }
+#endif
